simplify Booleanfilter and make it more generic
